requirements.txt
vllm>=0.6.0
torch>=2.0.0
transformers>=4.40.0
pyyaml>=6.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pydantic>=2.0.0
